Refactor to right package after refactoring.
